refactor(fb15b): extract BedrockStreamParser module
Move the colocated BedrockStreamParser class + its stream-type
interfaces out of bedrock.ts (1211->980 lines). 38 bedrock tests pass.
